refactor: 사용하지 않는 Fetcher.ts를 제거하고 공통 사용 코드를 api/common으로 이동
