+++ app/api/libraries/[libraryId]/stocks/loan/route.ts
@@ -1,4 +1,4 @@
-import apiClient from "@/utils/api/ApiClient";
+import apiClient from "@/utils/api/common/ApiClient";
 import { NextRequest, NextResponse } from "next/server";
 
 export const GET = async (req: NextRequest) => {
